+++ src/processor/scheduler.rs
@@ -15,7 +15,7 @@
 
 use std::collections::{BinaryHeap, HashMap, HashSet};
 use std::time::Instant;
-use std::sync::{mpsc, Arc};
+use std::sync::{mpsc, Arc, RwLock};
 
 use common::prelude::*;
 use common::state::{State, UniqueId};
@@ -50,7 +50,6 @@ impl<S: ScriptsRepositoryTrait> DebugDetails<S> {
 }
 
 
-#[derive(Clone)]
 pub enum SchedulerInput<S: ScriptsRepositoryTrait> {
     Job(Job<S>, isize),
     HealthStatus(mpsc::Sender<HealthDetails>),
@@ -63,6 +62,9 @@ pub enum SchedulerInput<S: ScriptsRepositoryTrait> {
     Lock,
     Unlock,
 
+    UpdateContext(JobContext<S>),
+    SetThreadsCount(u16),
+
     StopSignal,
     JobEnded(ScriptId<S>, ThreadCompleter),
 }
@@ -72,7 +74,7 @@ pub enum SchedulerInput<S: ScriptsRepositoryTrait> {
 pub struct Scheduler<S: ScriptsRepositoryTrait + 'static> {
     max_threads: u16,
     hooks: Arc<S>,
-    jobs_context: Arc<JobContext<S>>,
+    jobs_context: Arc<RwLock<Arc<JobContext<S>>>>,
     state: Arc<State>,
 
     locked: bool,
@@ -91,7 +93,7 @@ impl<S: ScriptsRepositoryTrait> Scheduler<S> {
     pub fn new(
         max_threads: u16,
         hooks: Arc<S>,
-        ctx: Arc<JobContext<S>>,
+        ctx: JobContext<S>,
         state: Arc<State>,
     ) -> Self {
         let (input_send, input_recv) = mpsc::channel();
@@ -107,7 +109,7 @@ impl<S: ScriptsRepositoryTrait> Scheduler<S> {
         Scheduler {
             max_threads: max_threads,
             hooks: hooks,
-            jobs_context: ctx,
+            jobs_context: Arc::new(RwLock::new(Arc::new(ctx))),
             state: state,
 
             locked: false,
@@ -209,8 +211,32 @@ impl<S: ScriptsRepositoryTrait> Scheduler<S> {
                     self.run_jobs();
                 }
 
+                SchedulerInput::UpdateContext(ctx) => {
+                    let mut ptr = self.jobs_context.write().unwrap();
+                    *ptr = Arc::new(ctx);
+                }
+
+                SchedulerInput::SetThreadsCount(max) => {
+                    self.max_threads = max;
+
+                    // Spawn new threads if the new maximum is higher, else
+                    // start cleaning up old ones
+                    if self.max_threads as usize > self.threads.len() {
+                        for _ in self.threads.len()..self.max_threads as usize {
+                            self.spawn_thread();
+                        }
+                    } else {
+                        self.cleanup_threads();
+                    }
+                }
+
                 SchedulerInput::JobEnded(hook_id, completer) => {
                     completer.manual_complete();
+
+                    // Cleanup threads if there are more than enough
+                    if self.threads.len() > self.max_threads as usize {
+                        self.cleanup_threads();
+                    }
 
                     // Put the highest-priority waiting job for this hook
                     // back in the queue
@@ -249,13 +275,14 @@ impl<S: ScriptsRepositoryTrait> Scheduler<S> {
 
     #[inline]
     fn spawn_thread(&mut self) {
-        let ctx = self.jobs_context.clone();
+        let ctx_lock = self.jobs_context.clone();
         let input = self.input_send.clone();
 
         let thread = Thread::new(
             move |job: ScheduledJob<S>, mut completer| {
                 completer.manual_mode();
 
+                let ctx = ctx_lock.read().unwrap().clone();
                 let result = job.execute(&ctx);
 
                 match result {
@@ -433,7 +460,7 @@ mod tests {
             let repo = Arc::new(Repository::<()>::new());
 
             let processor =
-                Processor::new(1, repo, Arc::new(()), Arc::new(State::new()))
+                Processor::new(1, repo, (), Arc::new(State::new()))
                     .unwrap();
             processor.stop()?;
 
@@ -457,7 +484,7 @@ mod tests {
             let processor = Processor::new(
                 1,
                 repo.clone(),
-                Arc::new(()),
+                (),
                 Arc::new(State::new()),
             )?;
 
@@ -491,7 +518,7 @@ mod tests {
         let processor = Processor::new(
             threads,
             repo.clone(),
-            Arc::new(()),
+            (),
             Arc::new(State::new()),
         )?;
 
@@ -561,7 +588,7 @@ mod tests {
             let processor = Processor::new(
                 2,
                 repo.clone(),
-                Arc::new(()),
+                (),
                 Arc::new(State::new()),
             )?;
             let api = processor.api();
@@ -625,7 +652,7 @@ mod tests {
             let processor = Processor::new(
                 1,
                 repo.clone(),
-                Arc::new(()),
+                (),
                 Arc::new(State::new()),
             )?;
             let api = processor.api();
@@ -675,7 +702,7 @@ mod tests {
             let processor = Processor::new(
                 1,
                 repo.clone(),
-                Arc::new(()),
+                (),
                 Arc::new(State::new()),
             )?;
             let api = processor.api();