+++ src/scripts/mod.rs
@@ -21,6 +21,6 @@ mod repository;
 mod script;
 
 pub use self::repository::{Blueprint, Repository};
-pub use self::repository::{ScriptNamesIter, ScriptsIter, StatusJobsIter};
+pub use self::repository::{ScriptsIter, StatusJobsIter};
 pub use self::script::{Script, ScriptProvider};
 pub use self::jobs::{Job, JobOutput, Context as JobContext};